Move _ListNode to its own file
